+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changes for Lovely Pytest Docker
 unreleased
 ==========
 
+ - fixed wrong error handling when polling for docker up state
+
 2018/01/10 0.0.2
 ================